Refactor [#23] Dto 네이밍 변경
