+++ morib/src/main/java/org/morib/server/api/homeView/vo/CategoryInfo.java
@@ -1,17 +1,17 @@
-package org.morib.server.api.homeView.dto.fetch;
+package org.morib.server.api.homeView.vo;
 
 import org.morib.server.domain.category.infra.Category;
 
 import java.time.LocalDate;
 
-public record FetchCategoryDto(
+public record CategoryInfo(
         Long id,
         String name,
         LocalDate startDate,
         LocalDate endDate
 ) {
-    public static FetchCategoryDto of(Category category) {
-        return new FetchCategoryDto(
+    public static CategoryInfo of(Category category) {
+        return new CategoryInfo(
                 category.getId(),
                 category.getName(),
                 category.getStartDate(),
