+++ morib/src/main/java/org/morib/server/api/homeView/vo/CombinedCategoryAndTaskInfo.java
@@ -1,0 +1,15 @@
+package org.morib.server.api.homeView.vo;
+
+import java.util.List;
+
+public record CombinedCategoryAndTaskInfo(
+        CategoryInfo category,
+        List<TaskInfo> tasks
+) {
+    public static CombinedCategoryAndTaskInfo of(CategoryInfo category, List<TaskInfo> tasks) {
+        return new CombinedCategoryAndTaskInfo(category, tasks);
+    }
+
+
+}
+
